+++ TODO.md
@@ -1,7 +1,21 @@
 # TODO
 
-[] Optimize timeline (https://www.w3schools.com/howto/howto_css_timeline.asp)
+### v0.2
+
+[X] Optimize timeline
+
 [] Change favicon
-[] Setup development branch
-[] Setup API endpoints for `About Me`
+
+[] Finalize design of `jobs/[id]` page
+
+### v0.3
+
+[] Setup API endpoints for `About Me`: `/api/about`
+
 [] Use above API endpoint to get content for `About Me` section
+
+[] Use `ThisIsJake.png` in header (left side)
+
+### Future plans/possible features
+
+[] Typescript version of About Me where text gets "typed in"
